requirements.txt
curl_cffi
fake_useragent
python-dotenv
Telethon
aiohttp
browser_cookie3
pillow
numpy